requirements.txt
# ---- web framework ----
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2

# ---- database ----
psycopg2-binary==2.9.11
SQLAlchemy==2.0.35

# ---- vector store ----
chromadb==0.5.20

# ---- embeddings ----
sentence-transformers==3.2.1
torch>=2.2.0

# ---- sparse retrieval ----
rank-bm25==0.2.2

# ---- LLM client (OpenRouter is OpenAI-compatible) ----
openai==1.54.4
httpx==0.27.2

# ---- math & utils ----
numpy==1.26.4
tenacity==9.0.0           # retries on LLM calls
python-dateutil==2.9.0

# ---- logging ----
structlog==24.4.0

# ---- testing ----
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
